+++ src/test/groovy/net/saliman/gradle/plugin/properties/PropertiesPluginParentProjectTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2016 Steven C. Saliman
+ * Copyright 2012-2019 Steven C. Saliman
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,6 +15,15 @@
  *
  */
 package net.saliman.gradle.plugin.properties
+
+import org.junit.Before
+import org.junit.Test
+
+import static org.junit.Assert.assertEquals
+import static org.junit.Assert.assertFalse
+import static org.junit.Assert.assertNull
+import static org.junit.Assert.fail
+
 /**
  * Test class for parent projects that apply Properties plugin.  In a multi
  * project build, parent and child projects use slightly different files.  A
@@ -33,7 +42,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * Set up the test data.  This calls a helper method to create the projects
 	 * because we need to repeat the setup in one of the tests.
 	 */
-	public void setUp() {
+	@Before
+	void setUp() {
 		createProjects()
 		setFileProperties(true, true)
 		copyFiles()
@@ -42,7 +52,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	/**
 	 * Test the CheckProperty method when the property is present.
 	 */
-	public void testCheckPropertyPresent() {
+	@Test
+	void checkPropertyPresent() {
 		parentProject.ext.someProperty = 'someValue'
 		// we succeed if we don't get an exception.
 		plugin.checkProperty(parentProject, 'someProperty', parentTask, 'someMethod')
@@ -53,12 +64,9 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	/**
 	 * Test the checkProperty method when the property is missing.
 	 */
-	public void testCheckPropertyMissing() {
-		// we succeed if we don't get an exception.
-		shouldFail(MissingPropertyException) {
-			plugin.checkProperty(parentProject, 'someProperty', parentTask, 'someMethod')
-		}
-
+	@Test(expected = MissingPropertyException.class)
+	void checkPropertyMissing() {
+		plugin.checkProperty(parentProject, 'someProperty', parentTask, 'someMethod')
 	}
 
 	// These tests are split up into multiples so that if one part works but
@@ -71,7 +79,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * .p.r.o.p.e.r.t.y_.n.a.m.e).  This test makes sure we don't do that any
 	 * more.
 	 */
-	public void testApplyUpperCommandProperty() {
+	@Test
+	void applyUpperCommandProperty() {
 		// simulate a "-PcommandProperty=Command.commandValue -PgradleUserName=user"
 		// command line
 		def commandArgs = [
@@ -98,7 +107,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * Verify that a command line value overrides everything else, and that we
 	 * don't set system properties from it.
 	 */
-	public void testApplyCommandProperty() {
+	@Test
+	void applyCommandProperty() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PsystemProp.commandProp=commandValue -PgradleUserName=user" command
 		// line
@@ -132,7 +142,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * line, system properties, or environment variables, the user value wins,
 	 * since it is the highest priority file.
 	 */
-	public void testApplyUserProperty() {
+	@Test
+	void applyUserProperty() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PsystemProp.commandProp=commandValue -PgradleUserName=user" command
 		// line
@@ -163,7 +174,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * Verify that when a file-based property is set in all files, but no user is
 	 * given, we get the value from the home file.
 	 */
-	public void testApplyUserPropertyNoUser() {
+	@Test
+	void applyUserPropertyNoUser() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PsystemProp.commandProp=commandValue" command line
 		def commandArgs = [
@@ -192,7 +204,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * Verify that when a file-based property is set everywhere but the user
 	 * file, the home file wins.
 	 */
-	public void testApplyHomeProperty() {
+	@Test
+	void applyHomeProperty() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PsystemProp.commandProp=commandValue -PgradleUserName=user" command
 		// line
@@ -226,7 +239,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * This also makes sure the files in the child project are ignored when we
 	 * apply at a parent level.
 	 */
-	public void testApplyUseDefaultFile() {
+	@Test
+	void applyUseDefaultFile() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PsystemProp.commandProp=commandValue -PgradleUserName=user" command
 		// line
@@ -269,7 +283,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * and not the one at the project level. This also makes sure the files in
 	 * the child project are ignored when we apply at a parent level.
 	 */
-	public void testApplyUseDefaultFileInDirectory() {
+	@Test
+	void applyUseDefaultFileInDirectory() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PsystemProp.commandProp=commandValue -PgradleUserName=user
 		// -PenvironmentFileName=gradle-properties" command line
@@ -311,7 +326,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * and we do specify an environment, the specified environment file wins.
 	 * This also verifies that properties in child project files are ignored.
 	 */
-	public void testApplyUseAlternateFile() {
+	@Test
+	void applyUseAlternateFile() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PsystemProp.commandProp=commandValue -PenvironmentName=test
 		// -PgradleUserName=user" command line
@@ -352,7 +368,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * it still gets set and is in the filters.  This also verifies that we
 	 * ignore properties set in child project files.
 	 */
-	public void testApplyProjectProperties() {
+	@Test
+	void applyProjectProperties() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PsystemProp.commandProp=commandValue -PgradleUserName=user"
 		// command line
@@ -395,7 +412,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * a user.  This is not an error.  Remember that child files should not be
 	 * processed.
 	 */
-	public void testApplyMissingUnspecifiedUserFile() {
+	@Test
+	void applyMissingUnspecifiedUserFile() {
 		def propFile = new File("${parentProject.gradle.gradleUserHomeDir}/gradle-user.properties")
 		propFile.delete()
 		assertFalse('Failed to delete user file', propFile.exists())
@@ -460,7 +478,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * Test applying the plugin when we have no user file for a specified user.
 	 * This should produce an error
 	 */
-	public void testApplyMissingSpecifiedUserFile() {
+	@Test
+	void applyMissingSpecifiedUserFile() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PgradleUserName=dummy" command line
 		def commandArgs = [
@@ -481,7 +500,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * Test what happens when we have no home file.  This should not produce an
 	 * error.
 	 */
-	public void testApplyMissingHomeFile() {
+	@Test
+	void applyMissingHomeFile() {
 		// delete the home file
 		def propFile = new File("${parentProject.gradle.gradleUserHomeDir}/gradle.properties")
 		propFile.delete()
@@ -555,7 +575,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * Test what happens when we have no environment file, and we're using the
 	 * default "local" file.  This should not be an error.
 	 */
-	public void testApplyMissingUnspecifiedEnvFile() {
+	@Test
+	void applyMissingUnspecifiedEnvFile() {
 		// delete the local file.
 		def propFile = new File("${parentProject.projectDir}/gradle-local.properties")
 		propFile.delete()
@@ -625,7 +646,8 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	 * Test what happens when we have no environment file, but we specify an
 	 * environment file.  This should be an error.
 	 */
-	public void testApplyMissingSpecifiedEnvFile() {
+	@Test
+	void applyMissingSpecifiedEnvFile() {
 		// simulate a "-PcommandProperty=Command.commandValue
 		// -PenvironmentName=dummy -PgradleUserName=user" command line
 		def commandArgs = [
@@ -644,9 +666,82 @@ class PropertiesPluginParentProjectTest extends BasePluginTest {
 	}
 
 	/**
+	 * Test what happens when we have no environment file and we specify an
+	 * environment file, but we tell the plugin to ignore missing files.  This
+	 * should behave the same as a missing local environment file; our
+	 * "environment" properties should come from the "project" file.
+	 */
+	@Test
+	void applyIgnoreMissingSpecifiedEnvFile() {
+		// simulate a "-PpropertiesPluginIgnoreMissingEnvFile=true
+		// -PcommandProperty=Command.commandValue
+		// -PenvironmentName=dummy -PgradleUserName=user" command line
+		def commandArgs = [
+				        propertiesPluginIgnoreMissingEnvFile: 'true',
+						commandProperty: 'Command.commandValue',
+						environmentName: 'dummy',
+						gradleUserName: 'user'
+		]
+		setNonFileProperties(true, true, commandArgs)
+
+		parentProject.apply plugin: 'properties'
+		def tokens = parentProject.filterTokens
+		assertEquals('dummy', parentProject.environmentName)
+		assertEquals('dummy', parentProject.ext.environmentName)
+		assertEquals('user', parentProject.gradleUserName)
+
+		assertEquals('ParentProject.parentProjectValue', parentProject.parentProjectProperty)
+		assertEquals('ParentProject.parentEnvironmentValue', parentProject.parentEnvironmentProperty)
+		assertEquals('ParentProject.childProjectValue', parentProject.childProjectProperty)
+		assertEquals('ParentProject.childEnvironmentValue', parentProject.childEnvironmentProperty)
+		assertEquals('Home.homeValue', parentProject.homeProperty)
+		assertEquals('User.userValue', parentProject.userProperty)
+		assertEquals('Environment.environmentValue', parentProject.environmentProperty)
+		assertEquals('System.systemValue', parentProject.systemProperty)
+		assertEquals('Command.commandValue', parentProject.commandProperty)
+
+		// 22 that we're checking, 14 from system properties, and 2 from our
+		// setting of the override property.
+		assertEquals(38, tokens.size())
+		// camel case notation
+		assertEquals('user', tokens['gradleUserName'])
+		assertEquals('ParentProject.parentProjectValue', tokens['parentProjectProperty'])
+		assertEquals('ParentProject.parentEnvironmentValue', tokens['parentEnvironmentProperty'])
+		assertEquals('ParentProject.childProjectValue', tokens['childProjectProperty'])
+		assertEquals('ParentProject.childEnvironmentValue', tokens['childEnvironmentProperty'])
+		assertEquals('Home.homeValue', tokens['homeProperty'])
+		assertEquals('User.userValue', tokens['userProperty'])
+		assertEquals('Environment.environmentValue', tokens['environmentProperty'])
+		assertEquals('System.systemValue', tokens['systemProperty'])
+		assertEquals('Command.commandValue', tokens['commandProperty'])
+		// dot notation
+		assertEquals('user', tokens['gradle.user.name'])
+		assertEquals('ParentProject.parentProjectValue', tokens['parent.project.property'])
+		assertEquals('ParentProject.parentEnvironmentValue', tokens['parent.environment.property'])
+		assertEquals('ParentProject.childProjectValue', tokens['child.project.property'])
+		assertEquals('ParentProject.childEnvironmentValue', tokens['child.environment.property'])
+		assertEquals('Home.homeValue', tokens['home.property'])
+		assertEquals('User.userValue', tokens['user.property'])
+		assertEquals('Environment.environmentValue', tokens['environment.property'])
+		assertEquals('System.systemValue', tokens['system.property'])
+		assertEquals('Command.commandValue', tokens['command.property'])
+
+		// Check the system properties. Remember, system properties are not
+		// set from command line properties.
+		assertEquals('ParentProject.parentProjectValue', System.properties['parentProjectProp'])
+		assertEquals('ParentProject.parentEnvironmentValue', System.properties['parentEnvironmentProp'])
+		assertEquals('ParentProject.childProjectValue', System.properties['childProjectProp'])
+		assertEquals('ParentProject.childEnvironmentValue', System.properties['childEnvironmentProp'])
+		assertEquals('Home.homeValue', System.properties['homeProp'])
+		assertEquals('User.userValue', System.properties['userProp'])
+		assertEquals('User.commandValue', System.properties['commandProp'])
+	}
+
+	/**
 	 * Test what happens when we have no project property file.  This is no error.
 	 */
-	public void testApplyMissingProjectFile() {
+	@Test
+	void applyMissingProjectFile() {
 		def propFile = new File("${parentProject.projectDir}/gradle.properties")
 		propFile.delete()
 		assertFalse('Failed to delete project file', propFile.exists())